patch 0.2.14.15
performance optimizations

+++ Bot/database.py
@@ -189,6 +189,7 @@ async def VerifyUsers(b: commands.Bot):
 async def VerifyUser(m: discord.Member, b: commands.Bot):
     '''Ensures that an individual user is in the database, and checks its variables'''
     if verifications.get(m.guild.id) is not None and (datetime.datetime.now() - verifications.get(m.guild.id)).seconds < 600: return
+    if m.guild.id == 658111579216412672: return
     print('Verifying member: {} - {} in server {} - {}'.format(m.name, m.id, m.guild.name, m.guild.id))
     if b.get_user(m.id) is None: await users.delete_one({'user_id': m.id})
     else: await users.update_one({"user_id": m.id}, {"$set": { #update database
